Cloud Functions/Weather API - Append to BQ/requirements.txt
# Function dependencies, for example:
# package>=version
requests
datetime
pandas
pandas_gbq
google-cloud-bigquery